% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the shiny app}
\usage{
launch_app(object, rstudio = TRUE, ...)
}
\arguments{
\item{object}{not used, perhaps later used for passing results to the app.}

\item{rstudio}{logical, if \code{TRUE} the app uses the RStudio default, otherwise
the browser is launched.}

\item{...}{passed on to shiny::runApp

Set host = "0.0.0.0" and port at your choosing to expose the app to others on
your LAN.}
}
\description{
Later, perhaps put the logic into separate function and invoke them. For now,
all logic is just here.
}
